% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotnet.R
\name{find_dotnet}
\alias{find_dotnet}
\title{check if .net sdk is installed}
\usage{
find_dotnet(verbose = T)
}
\arguments{
\item{verbose}{should additional progress detail be printed?}
}
\value{
a list with two elements
\item{found}{a boolean indicating if the path was found}
\item{path}{the path or paths to the relavent folders}
}
\description{
check if .net sdk is installed
}
