% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{to_fsi}
\alias{to_fsi}
\title{Evaluate Code in FSI}
\usage{
to_fsi(code, terminal = NA, show = T)
}
\arguments{
\item{code}{Fsharp code to be executed in FSI}

\item{terminal}{relevant terminal id}

\item{show}{should it bring the terminal window to focus?}
}
\description{
Evaluate Code in FSI
}
